+++ HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds the ability to define a success and an error url.
+  [href]
+
 0.2.0 (2017-05-24)
 ~~~~~~~~~~~~~~~~~~~~~